% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_calibration}
\alias{plot_calibration}
\title{Plot calibration curve}
\usage{
plot_calibration(observed, predicted, groups = 10)
}
\arguments{
\item{observed}{Binary observed outcome, coded 0/1.}

\item{predicted}{Predicted probabilities.}

\item{groups}{Number of quantile groups.}
}
\value{
A ggplot object.
}
\description{
Plot calibration curve
}
