% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_risk_groups}
\alias{plot_risk_groups}
\title{Plot risk group event rates}
\usage{
plot_risk_groups(observed, risk_group)
}
\arguments{
\item{observed}{Binary observed outcome, coded 0/1.}

\item{risk_group}{Ordered risk group factor.}
}
\value{
A ggplot object.
}
\description{
Plot risk group event rates
}
